fix event author unexpected reset
calling `ndk.getUser` will always return a new instance of `NDKUser` which will cause a reset in author object every time we call `event.author`. this behavior is not expected!
for example after calling `event.author.fetchProfile()` we expect that the `event.author` object is now containing the `profile` object, but we are still getting `undefined` though we've already fetched the `profile`!
so I've added a private `_author` property inside each event instance, which is responsible for holding the actual value of `author`.
now we don't get a newly created partial NDKUser every time we call `event.author`.