update svelte-asciidoc.
this fixes an issue with tables.

it also introduces a way to disable the markdown transitional syntax
that makes the asciidoc parser understand some random markdown syntax
directives like "#" (which should be "=" in asciidoc) and just
confuses everybody.

it should have been disabled since day one, but since I missed that
this commit keeps it for all articles in the past and disables it only
for events authored after 2024-09-14 (two weeks expected for this change
to be deployed).